fix: filter out commands/builtins starting with ./
