+++ include/execution.h
@@ -6,7 +6,7 @@
 /*   By: aimokhta <aimokhta@student.42kl.edu.my>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/15 18:40:52 by aimokhta          #+#    #+#             */
-/*   Updated: 2025/07/24 09:47:01 by aimokhta         ###   ########.fr       */
+/*   Updated: 2025/07/24 11:46:22 by aimokhta         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -83,6 +83,7 @@ void	rd_in(t_ast *ast, t_exc *exc);
 void	rd_out(t_ast *ast, t_exc *exc);
 void	rd_heredoc(t_ast *ast, t_exc *exc);
 void	exec_word(t_ast *ast, t_exc *exc);
+void	access_and_execve(t_exc *exc, t_ast *ast);;
 void	dupping_stdin_stdout(t_exc *exc);
 void	dup2_close_infile_outfile(t_exc *exc);
 void	close_infile_outfile_parent(t_exc *exc);
