+++ src/execution/exec/word/built_in/built_ins.c
@@ -6,7 +6,7 @@
 /*   By: aimokhta <aimokhta@student.42kl.edu.my>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/15 18:15:50 by aimokhta          #+#    #+#             */
-/*   Updated: 2025/07/22 10:45:41 by aimokhta         ###   ########.fr       */
+/*   Updated: 2025/07/24 11:47:57 by aimokhta         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,6 +18,25 @@
 // but then come back to the original room remain unchanged
 // other than builtins - called external programs (ls, cat, etc)
 
+int	is_bi(char **args)
+{
+	if (ft_strncmp(args[0], "echo", 5) == 0)
+		return (1);
+	else if (ft_strncmp(args[0], "pwd", 4) == 0)
+		return (1);
+	else if (ft_strncmp(args[0], "env", 4) == 0)
+		return (1);
+	else if (ft_strncmp(args[0], "export", 7) == 0)
+		return (1);
+	else if (ft_strncmp(args[0], "cd", 3) == 0)
+		return (1);
+	else if (ft_strncmp(args[0], "unset", 6) == 0)
+		return (1);
+	else if (ft_strncmp(args[0], "exit", 5) == 0)
+		return (1);
+	return (0);
+}
+
 void	exec_builtin(t_ast *ast, t_exc *exc)
 {
 	dupping_stdin_stdout(exc);
@@ -26,11 +45,12 @@ void	exec_builtin(t_ast *ast, t_exc *exc)
 	reset_stdin_stdout_unlink_heredocfd(exc);
 }
 
+// for 
 void	built_ins(char **args, t_exc *exc)
 {
-	if (!args[0])
-		return ;
-	else if (ft_strncmp(args[0], "echo", 5) == 0)
+	// if (!args[0])
+	// 	return ;
+	if (ft_strncmp(args[0], "echo", 5) == 0)
 		bi_echo(args, exc);
 	else if (ft_strncmp(args[0], "cd", 3) == 0)
 		bi_cd(args, exc);
@@ -45,5 +65,8 @@ void	built_ins(char **args, t_exc *exc)
 	else if (ft_strncmp(args[0], "exit", 5) == 0)
 		bi_exit(args, exc);
 	else
+	{
+		fprintf(stderr, "args is empty\n");
 		return ;
+	}
 }
