+++ src/execution/exec/word/command/exec_word.c
@@ -6,16 +6,13 @@
 /*   By: aimokhta <aimokhta@student.42kl.edu.my>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/17 10:07:12 by aimokhta          #+#    #+#             */
-/*   Updated: 2025/07/21 17:07:25 by aimokhta         ###   ########.fr       */
+/*   Updated: 2025/07/24 12:03:11 by aimokhta         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "execution.h"
 
 void		exec_word(t_ast *ast, t_exc *exc);
-static void	access_and_execve(t_exc *exc, t_ast *ast);
-static char	*access_path(t_exc *exc, char **args);
-static void	pathname_error_handling(t_exc *exc, char *pathname, char **args);
 static void	waiting_exec_word(pid_t pid, int exit_status, t_exc *exc);
 
 void	exec_word(t_ast *ast, t_exc *exc)
@@ -44,73 +41,6 @@ void	exec_word(t_ast *ast, t_exc *exc)
 		close_infile_outfile_parent(exc);
 		waiting_exec_word(pid, exit_status, exc);
 		reset_stdin_stdout_unlink_heredocfd(exc);
-	}
-}
-
-static void	access_and_execve(t_exc *exc, t_ast *ast)
-{
-	char	**args;
-	char	*pathname;
-
-	args = ast->token->basin_buff;
-	reset_signals_child();
-	if (!args || !args[0] || !args[0][0])
-	{
-		free_before_readline(exc);
-		exc->exit_code = CMD_NOT_FOUND;
-		exit(exc->exit_code);
-	}
-	if (access(args[0], F_OK) != -1)
-		pathname = ft_strdup(args[0]);
-	else
-		pathname = access_path(exc, args);
-	pathname_error_handling(exc, pathname, args);
-	execve(pathname, args, exc->exec->envp_array);
-	free_before_readline(exc);
-	free(pathname);
-	exc->exit_code = CMD_NOT_FOUND;
-	exit(exc->exit_code);
-}
-
-///usr/bin is a directory
-// that stores many executable programs (binaries) like ls/cat/grep/etc
-//
-// /usr/bin/ls is a binary executable file
-// we join them to fund where is the executable ls is stored
-// when you run ls, you are running the program located at /usr/bin/ls
-//
-// access - to see if the normal file (/usr/bin/ls) exists 
-// a.k.a. executable binary file
-static char	*access_path(t_exc *exc, char **args)
-{
-	int		i;
-	char	*pathname;
-
-	i = 0;
-	while (exc->process->splitted_path[i])
-	{
-		pathname = ft_strjoin(exc->process->splitted_path[i], args[0]);
-		if (access(pathname, F_OK) == 0)
-			return (pathname);
-		else
-		{
-			free(pathname);
-			i++;
-		}
-	}
-	return (NULL);
-}
-
-static void	pathname_error_handling(t_exc *exc, char *pathname, char **args)
-{
-	if (!pathname || !pathname[0])
-	{
-		ft_putstr_fd("shellfish: command not found: ", 2);
-		ft_putendl_fd(args[0], 2);
-		free_before_readline(exc);
-		free(pathname);
-		exc->exit_code = CMD_NOT_FOUND;
-		exit(exc->exit_code);
 	}
 }
